Add KindOf to AST
NB: There’s probably a better way of representing this whole thing than
a singly linked list, which I can only assume is going to come back and
bite us in the ass when it comes time to implement kind checking.

Needed to use the identifier “KindOf” instead of “Kind” because too
many things are named “Kind” around here and I don’t need to go
stepping on any toes.